thicc reorg
