+++ src/logger/mod.rs
@@ -1,35 +1,9 @@
-use std::any::Any;
+use crate::worlds::{Event, State};
 use std::collections::{BTreeMap, BinaryHeap};
 
-use crate::worlds::{Event, State};
+mod snapshot;
 
-/// A snapshot of the world at a given time.
-#[derive(Clone)]
-pub struct Snapshot {
-    pub timestamp: f64,
-    pub shared_state: Option<State>,
-    pub agent_states: BTreeMap<usize, State>,
-}
-
-impl PartialEq for Snapshot {
-    fn eq(&self, other: &Self) -> bool {
-        self.timestamp == other.timestamp
-    }
-}
-
-impl Eq for Snapshot {}
-
-impl PartialOrd for Snapshot {
-    fn partial_cmp(&self, other: &Self) -> Option<std::cmp::Ordering> {
-        Some(self.cmp(other))
-    }
-}
-
-impl Ord for Snapshot {
-    fn cmp(&self, other: &Self) -> std::cmp::Ordering {
-        self.timestamp.partial_cmp(&other.timestamp).unwrap()
-    }
-}
+pub use snapshot::Snapshot;
 
 /// A logger for recording snapshots of the world.
 pub struct Logger {
