+++ src/worlds/error.rs
@@ -1,0 +1,13 @@
+/// Error enum for provide feedback on simulation errors
+#[derive(Debug, Clone)]
+pub enum SimError {
+    TimeTravel,
+    PastTerminal,
+    ScheduleFailed,
+    PlaybackFroze,
+    NoState,
+    NoEvents,
+    NoClock,
+    InvalidIndex,
+    TokioError(String),
+}
